move llm interface
